Backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.1
pymongo==4.16.0
python-dotenv==1.1.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pydantic[email]==2.11.1
pydantic-settings==2.8.1
groq>=0.11.0
